feat: rename dashboard dynamic routing
